ai-service/requirements.txt
google-genai
python-dotenv
fastapi
uvicorn